# Conflicts:
#	.gitignore

+++ controller/midjourney.go
@@ -175,7 +175,7 @@ func UpdateMidjourneyTaskBulk() {
 				if err != nil {
 					logger.LogError(ctx, "UpdateMidjourneyTask task error: "+err.Error())
 				} else if won && shouldReturnQuota {
-					err = model.IncreaseUserQuota(task.UserId, task.Quota, false)
+					err = model.IncreaseUserQuota(task.UserId, int64(task.Quota), false)
 					if err != nil {
 						logger.LogError(ctx, "fail to increase user quota: "+err.Error())
 					}
